requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
django_cors_headers==4.3.0
